pyproject.toml
[project]
name = "citationManager"
version = "0.0.1"
description = "A citation manager for the diSimplex projects which understands our NikolaBase citation system"
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "pybtex>=0.24.0",
    "rispy>=0.8.1",
    "pybtexris>=0.1.2",
    "nicegui>=1.3.14",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
cmScan    = "citationManager.scanner:cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.20.3",
    "pytest-mock>=3.10.0",
]
